+++ libevmasm/GasMeter.h
@@ -95,6 +95,7 @@ namespace GasCosts
 	static unsigned const txDataNonZeroGas = 68;
 	static unsigned const copyGas = 3;
 	static unsigned const isValidatorGas = 3;
+	static unsigned const freeGasGas = 3;
 }
 
 /**